Hice un refactor del proyecto para seguir trabajando
